move `init_prod` migrations into separate module
