+++ src/models/src/migration/db_migrate.rs
@@ -8,7 +8,7 @@ use crate::entity::colors::ColorEntity;
 use crate::entity::config::ConfigEntity;
 use crate::entity::devices::DeviceEntity;
 use crate::entity::groups::Group;
-use crate::entity::jwk::{Jwk, JwkKeyPairAlg};
+use crate::entity::jwk::Jwk;
 use crate::entity::logos::Logo;
 use crate::entity::magic_links::MagicLink;
 use crate::entity::password::RecentPasswordsEntity;
@@ -24,387 +24,12 @@ use crate::entity::webauthn::PasskeyEntity;
 use crate::entity::webids::WebId;
 use crate::events::event::Event;
 use crate::migration::inserts;
-use actix_web::web;
-use argon2::password_hash::SaltString;
-use argon2::{Algorithm, Argon2, Params, PasswordHasher, Version};
-use cryptr::{EncKeys, EncValue};
 use hiqlite::{params, Param};
-use jwt_simple::algorithms::{
-    Ed25519KeyPair, EdDSAKeyPairLike, RS256KeyPair, RS384KeyPair, RS512KeyPair, RSAKeyPairLike,
-};
-use rand_core::OsRng;
-use rauthy_api_types::api_keys::ApiKeyRequest;
-use rauthy_common::constants::{
-    ADMIN_FORCE_MFA, DB_TYPE, DEV_MODE, PUB_URL, PUB_URL_WITH_SCHEME, RAUTHY_ADMIN_EMAIL,
-};
-use rauthy_common::utils::{base64_decode, get_rand};
-use rauthy_common::{is_hiqlite, DbType};
+use rauthy_common::constants::DB_TYPE;
+use rauthy_common::DbType;
 use rauthy_error::ErrorResponse;
-use ring::digest;
-use sqlx::{query, Row};
-use std::env;
-use time::OffsetDateTime;
+use sqlx::Row;
 use tracing::{debug, info};
-use validator::Validate;
-
-pub async fn anti_lockout(issuer: &str) -> Result<(), ErrorResponse> {
-    debug!("Executing anti_lockout_check");
-
-    let (redirect_uris, allowed_origins) = if *DEV_MODE {
-        let (ip, _) = PUB_URL.split_once(':').expect("PUB_URL must have a port");
-        let origin = if ip != "localhost" {
-            format!("https://{}:5173", ip)
-        } else {
-            "http://localhost:5173".to_string()
-        };
-
-        (
-            format!(
-                "{issuer}/oidc/callback,http://localhost:5173/*,https://{}:5173/*",
-                ip
-            ),
-            Some(origin),
-        )
-    } else {
-        (format!("{issuer}/oidc/callback"), None)
-    };
-
-    // we will actually skip non-mandatory values in the query below
-    let rauthy = Client {
-        id: "rauthy".to_string(),
-        name: None,
-        enabled: true,
-        confidential: false,
-        secret: None,
-        secret_kid: None,
-        redirect_uris: redirect_uris.clone(),
-        post_logout_redirect_uris: Some(redirect_uris),
-        allowed_origins,
-        flows_enabled: "authorization_code".to_string(),
-        access_token_alg: "EdDSA".to_string(),
-        id_token_alg: "EdDSA".to_string(),
-        auth_code_lifetime: 10,
-        access_token_lifetime: 10, // The token is not used for the Admin UI -> session only
-        scopes: "openid".to_string(),
-        default_scopes: "openid".to_string(),
-        challenge: Some("S256".to_string()),
-        force_mfa: *ADMIN_FORCE_MFA,
-        client_uri: Some(PUB_URL_WITH_SCHEME.to_string()),
-        contacts: RAUTHY_ADMIN_EMAIL.clone(),
-    };
-
-    // MUST NOT use `insert or replace` syntax
-    // -> SQLite basically re-creates this under the hood, which means the FK restriction
-    // from `client_logos` -> `clients` will actually delete the client logo that was
-    // saved for `rauthy` before.
-    // Update only here and prevent `rauthy` deletion as a special check on DELETE /client
-
-    if is_hiqlite() {
-        DB::client()
-            .execute(
-                r#"
-UPDATE clients SET enabled = $1, confidential = $2, redirect_uris = $3,
-post_logout_redirect_uris = $4, allowed_origins = $5, flows_enabled = $6, access_token_alg = $7,
-id_token_alg = $8, auth_code_lifetime = $9, access_token_lifetime = $10, scopes = $11,
-default_scopes = $12, challenge = $13, force_mfa = $14, client_uri = $15, contacts = $16
-WHERE id = $17"#,
-                params!(
-                    rauthy.enabled,
-                    rauthy.confidential,
-                    rauthy.redirect_uris,
-                    rauthy.post_logout_redirect_uris,
-                    rauthy.allowed_origins,
-                    rauthy.flows_enabled,
-                    rauthy.access_token_alg,
-                    rauthy.id_token_alg,
-                    rauthy.auth_code_lifetime,
-                    rauthy.access_token_lifetime,
-                    rauthy.scopes,
-                    rauthy.default_scopes,
-                    rauthy.challenge,
-                    rauthy.force_mfa,
-                    rauthy.client_uri,
-                    rauthy.contacts,
-                    rauthy.id
-                ),
-            )
-            .await?;
-    } else {
-        sqlx::query!(
-            r#"
-UPDATE clients SET enabled = $1, confidential = $2, redirect_uris = $3,
-post_logout_redirect_uris = $4, allowed_origins = $5, flows_enabled = $6, access_token_alg = $7,
-id_token_alg = $8, auth_code_lifetime = $9, access_token_lifetime = $10, scopes = $11,
-default_scopes = $12, challenge = $13, force_mfa = $14, client_uri = $15, contacts = $16
-WHERE id = $17"#,
-            rauthy.enabled,
-            rauthy.confidential,
-            rauthy.redirect_uris,
-            rauthy.post_logout_redirect_uris,
-            rauthy.allowed_origins,
-            rauthy.flows_enabled,
-            rauthy.access_token_alg,
-            rauthy.id_token_alg,
-            rauthy.auth_code_lifetime,
-            rauthy.access_token_lifetime,
-            rauthy.scopes,
-            rauthy.default_scopes,
-            rauthy.challenge,
-            rauthy.force_mfa,
-            rauthy.client_uri,
-            rauthy.contacts,
-            rauthy.id,
-        )
-        .execute(DB::conn())
-        .await?;
-    }
-
-    Ok(())
-}
-
-/// Initializes an empty production database for a new deployment
-pub async fn migrate_init_prod(argon2_params: Params, issuer: &str) -> Result<(), ErrorResponse> {
-    // check if the database is un-initialized by looking at the jwks table, which should be empty
-    let jwks = if is_hiqlite() {
-        DB::client()
-            .query_as("SELECT * FROM JWKS", params!())
-            .await?
-    } else {
-        sqlx::query_as::<_, Jwk>("SELECT * FROM JWKS")
-            .fetch_all(DB::conn())
-            .await?
-    };
-
-    if jwks.is_empty() {
-        info!("Initializing empty production database");
-
-        // For initializing a prod database, we need to:
-        // - delete init_admin and client
-        // - set new random password for admin and log to console with the first startup
-        // - generate a new set of JWKs
-
-        // cleanup
-        if is_hiqlite() {
-            DB::client()
-                .execute("DELETE FROM clients WHERE id = 'init_client'", params!())
-                .await?;
-            DB::client().execute("DELETE FROM users WHERE email IN ('init_admin@localhost.de', 'test_admin@localhost.de')", params!())
-                .await?;
-        } else {
-            sqlx::query!("DELETE FROM clients WHERE id = 'init_client'")
-                .execute(DB::conn())
-                .await?;
-
-            sqlx::query!(
-                "DELETE FROM users WHERE email IN ('init_admin@localhost.de', 'test_admin@localhost.de')",
-            )
-            .execute(DB::conn())
-            .await?;
-        }
-
-        // check if we should use manually provided bootstrap values
-        let email =
-            env::var("BOOTSTRAP_ADMIN_EMAIL").unwrap_or_else(|_| "admin@localhost.de".to_string());
-        let hash = match env::var("BOOTSTRAP_ADMIN_PASSWORD_ARGON2ID") {
-            Ok(hash) => {
-                info!(
-                    r#"
-
-    First-Time setup - an already hashed bootstrap password has been given for '{email}'
-
-    Please change it immediately: {issuer}/account
-    You will never see this message again!
-        "#
-                );
-                hash
-            }
-            Err(_) => {
-                let plain = match env::var("BOOTSTRAP_ADMIN_PASSWORD_PLAIN") {
-                    Ok(plain) => {
-                        info!(
-                            r#"
-
-    First-Time setup - a bootstrap password has been given for '{email}'
-
-    Please change it immediately: {issuer}/account
-    You will never see this message again!
-        "#
-                        );
-                        plain
-                    }
-                    Err(_) => {
-                        let plain = get_rand(24);
-                        info!(
-                            r#"
-
-    First-Time setup - new random password for '{email}':
-
-    {plain}
-
-    Please change it immediately: {issuer}/account
-    You will never see this message again!
-        "#
-                        );
-                        plain
-                    }
-                };
-
-                let argon2 = Argon2::new(Algorithm::Argon2id, Version::V0x13, argon2_params);
-                let salt = SaltString::generate(&mut OsRng);
-                argon2
-                    .hash_password(plain.as_bytes(), &salt)
-                    .expect("Error hashing the Password")
-                    .to_string()
-            }
-        };
-
-        if is_hiqlite() {
-            // TODO we could grab a possibly existing `RAUTHY_ADMIN_EMAIL` and initialize a custom
-            // admin
-            DB::client()
-                .execute(
-                    "UPDATE users SET email = $1, password = $2 WHERE email = 'admin@localhost.de'",
-                    params!(email, hash),
-                )
-                .await?;
-        } else {
-            sqlx::query!(
-                "UPDATE users SET email = $1, password = $2 WHERE email = 'admin@localhost.de'",
-                email,
-                hash
-            )
-            .execute(DB::conn())
-            .await?;
-        }
-
-        // now check if we should bootstrap an API Key
-        if let Ok(api_key_raw) = env::var("BOOTSTRAP_API_KEY") {
-            // this is expected to be base64 encoded
-            let bytes = base64_decode(&api_key_raw)?;
-            // ... and then a JSON string
-            let s = String::from_utf8_lossy(&bytes);
-            let req = serde_json::from_str::<ApiKeyRequest>(&s)?;
-            req.validate()
-                .expect("Invalid API Key in BOOTSTRAP_API_KEY");
-
-            debug!("Bootstrapping API Key:\n{:?}", req);
-            let key_name = req.name.clone();
-            let _ = ApiKeyEntity::create(
-                req.name,
-                req.exp,
-                req.access.into_iter().map(|a| a.into()).collect(),
-            )
-            .await?;
-
-            if let Ok(secret_plain) = env::var("BOOTSTRAP_API_KEY_SECRET") {
-                assert!(secret_plain.len() >= 64);
-                let secret_hash = digest::digest(&digest::SHA256, secret_plain.as_bytes());
-                let secret_enc = EncValue::encrypt(secret_hash.as_ref())?
-                    .into_bytes()
-                    .to_vec();
-
-                if is_hiqlite() {
-                    DB::client()
-                        .execute(
-                            "UPDATE api_keys SET secret = $1 WHERE name = $2",
-                            params!(secret_enc, key_name),
-                        )
-                        .await?;
-                } else {
-                    query!(
-                        "UPDATE api_keys SET secret = $1 WHERE name = $2",
-                        secret_enc,
-                        key_name,
-                    )
-                    .execute(DB::conn())
-                    .await?;
-                }
-            }
-        }
-
-        let enc_key_active = &EncKeys::get_static().enc_key_active;
-
-        // generate JWKs
-        info!("Generating new JWKs - this might take a few seconds");
-        let mut entities = Vec::with_capacity(4);
-
-        // RSA256
-        let jwk_plain = web::block(|| {
-            RS256KeyPair::generate(2048)
-                .unwrap()
-                .with_key_id(&get_rand(24))
-        })
-        .await?;
-        let jwk = EncValue::encrypt(jwk_plain.to_der().unwrap().as_slice())?
-            .into_bytes()
-            .to_vec();
-        entities.push(Jwk {
-            kid: jwk_plain.key_id().as_ref().unwrap().clone(),
-            created_at: OffsetDateTime::now_utc().unix_timestamp(),
-            signature: JwkKeyPairAlg::RS256,
-            enc_key_id: enc_key_active.clone(),
-            jwk,
-        });
-
-        // RS384
-        let jwk_plain = web::block(|| {
-            RS384KeyPair::generate(3072)
-                .unwrap()
-                .with_key_id(&get_rand(24))
-        })
-        .await?;
-        let jwk = EncValue::encrypt(jwk_plain.to_der().unwrap().as_slice())?
-            .into_bytes()
-            .to_vec();
-        entities.push(Jwk {
-            kid: jwk_plain.key_id().as_ref().unwrap().clone(),
-            created_at: OffsetDateTime::now_utc().unix_timestamp(),
-            signature: JwkKeyPairAlg::RS384,
-            enc_key_id: enc_key_active.clone(),
-            jwk,
-        });
-
-        // RSA512
-        let jwk_plain = web::block(|| {
-            RS512KeyPair::generate(4096)
-                .unwrap()
-                .with_key_id(&get_rand(24))
-        })
-        .await?;
-        let jwk = EncValue::encrypt(jwk_plain.to_der().unwrap().as_slice())?
-            .into_bytes()
-            .to_vec();
-        entities.push(Jwk {
-            kid: jwk_plain.key_id().as_ref().unwrap().clone(),
-            created_at: OffsetDateTime::now_utc().unix_timestamp(),
-            signature: JwkKeyPairAlg::RS512,
-            enc_key_id: enc_key_active.clone(),
-            jwk,
-        });
-
-        // Ed25519
-        let jwk_plain =
-            web::block(|| Ed25519KeyPair::generate().with_key_id(&get_rand(24))).await?;
-        let jwk = EncValue::encrypt(jwk_plain.to_der().as_slice())?
-            .into_bytes()
-            .to_vec();
-        entities.push(Jwk {
-            kid: jwk_plain.key_id().as_ref().unwrap().clone(),
-            created_at: OffsetDateTime::now_utc().unix_timestamp(),
-            signature: JwkKeyPairAlg::EdDSA,
-            enc_key_id: enc_key_active.clone(),
-            jwk,
-        });
-
-        for e in entities {
-            e.save().await?;
-        }
-
-        info!("Production database initialized successfully");
-    }
-
-    Ok(())
-}
 
 // NOTE:
 // All the queries in the migrations here cannot be compile-time checked.
