+++ src/models/src/migration/mod.rs
@@ -18,8 +18,10 @@ use tokio::fs;
 use tokio::time::Instant;
 use tracing::{debug, error, info, warn};
 
+pub mod anti_lockout;
 pub mod db_migrate;
 pub mod db_migrate_dev;
+pub mod init_prod;
 mod inserts;
 
 static BUCKET: OnceLock<Bucket> = OnceLock::new();
